# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Movie & Book Review API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  username: string;
  email: string;
  createdAt: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type ReviewMediaType =
  (typeof ReviewMediaType)[keyof typeof ReviewMediaType];

export const ReviewMediaType = {
  movie: "movie",
  book: "book",
} as const;

export interface Review {
  id: number;
  userId: number;
  /** @nullable */
  username?: string | null;
  mediaType: ReviewMediaType;
  externalId: string;
  title: string;
  /** @nullable */
  coverImageUrl?: string | null;
  /** @nullable */
  year?: string | null;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  reviewText: string;
  createdAt: string;
}

export type ReviewInputMediaType =
  (typeof ReviewInputMediaType)[keyof typeof ReviewInputMediaType];

export const ReviewInputMediaType = {
  movie: "movie",
  book: "book",
} as const;

export interface ReviewInput {
  mediaType: ReviewInputMediaType;
  externalId: string;
  title: string;
  /** @nullable */
  coverImageUrl?: string | null;
  /** @nullable */
  year?: string | null;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  reviewText: string;
}

export interface ReviewsPage {
  reviews: Review[];
  total: number;
  page: number;
  limit: number;
}

export interface ReviewStats {
  totalReviews: number;
  totalMovieReviews: number;
  totalBookReviews: number;
  averageRating: number;
  topRated: Review[];
}

export type WatchlistItemMediaType =
  (typeof WatchlistItemMediaType)[keyof typeof WatchlistItemMediaType];

export const WatchlistItemMediaType = {
  movie: "movie",
  book: "book",
} as const;

export interface WatchlistItem {
  id: number;
  userId: number;
  mediaType: WatchlistItemMediaType;
  externalId: string;
  title: string;
  /** @nullable */
  coverImageUrl?: string | null;
  addedAt: string;
}

export type WatchlistInputMediaType =
  (typeof WatchlistInputMediaType)[keyof typeof WatchlistInputMediaType];

export const WatchlistInputMediaType = {
  movie: "movie",
  book: "book",
} as const;

export interface WatchlistInput {
  mediaType: WatchlistInputMediaType;
  externalId: string;
  title: string;
  /** @nullable */
  coverImageUrl?: string | null;
}

export type SearchResultMediaType =
  (typeof SearchResultMediaType)[keyof typeof SearchResultMediaType];

export const SearchResultMediaType = {
  movie: "movie",
  book: "book",
} as const;

export interface SearchResult {
  externalId: string;
  title: string;
  mediaType: SearchResultMediaType;
  /** @nullable */
  coverImageUrl?: string | null;
  /** @nullable */
  year?: string | null;
  /** @nullable */
  author?: string | null;
}

export type ListReviewsParams = {
  page?: number;
  limit?: number;
  mediaType?: ListReviewsMediaType;
};

export type ListReviewsMediaType =
  (typeof ListReviewsMediaType)[keyof typeof ListReviewsMediaType];

export const ListReviewsMediaType = {
  movie: "movie",
  book: "book",
} as const;

export type GetRecentReviewsParams = {
  limit?: number;
};

export type SearchParams = {
  q: string;
  type: SearchType;
};

export type SearchType = (typeof SearchType)[keyof typeof SearchType];

export const SearchType = {
  movie: "movie",
  book: "book",
} as const;