# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Movie & Book Review API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: reviews
    description: Review operations
  - name: watchlist
    description: Watchlist operations
  - name: search
    description: Search operations
  - name: stats
    description: Stats and aggregates

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List all reviews (paginated)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
        - name: mediaType
          in: query
          schema:
            type: string
            enum: [movie, book]
      responses:
        "200":
          description: Reviews list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewsPage"

    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "401":
          description: Unauthorized

  /reviews/recent:
    get:
      operationId: getRecentReviews
      tags: [reviews]
      summary: Get recent reviews (for home feed)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 6
      responses:
        "200":
          description: Recent reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

  /reviews/stats:
    get:
      operationId: getReviewStats
      tags: [stats]
      summary: Get review stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewStats"

  /reviews/{id}:
    get:
      operationId: getReview
      tags: [reviews]
      summary: Get a single review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "404":
          description: Not found

    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete own review
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /users/{userId}/reviews:
    get:
      operationId: getUserReviews
      tags: [reviews]
      summary: Get reviews by user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

  /watchlist:
    get:
      operationId: getWatchlist
      tags: [watchlist]
      summary: Get current user's watchlist
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Watchlist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WatchlistItem"
        "401":
          description: Unauthorized

    post:
      operationId: addToWatchlist
      tags: [watchlist]
      summary: Add to watchlist
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistItem"
        "401":
          description: Unauthorized

  /watchlist/{id}:
    delete:
      operationId: removeFromWatchlist
      tags: [watchlist]
      summary: Remove from watchlist
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
        "401":
          description: Unauthorized
        "404":
          description: Not found

  /search:
    get:
      operationId: search
      tags: [search]
      summary: Search movies or books
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [movie, book]
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SearchResult"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    User:
      type: object
      required: [id, username, email, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time

    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    Review:
      type: object
      required: [id, userId, mediaType, externalId, title, rating, reviewText, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: ["string", "null"]
        mediaType:
          type: string
          enum: [movie, book]
        externalId:
          type: string
        title:
          type: string
        coverImageUrl:
          type: ["string", "null"]
        year:
          type: ["string", "null"]
        rating:
          type: integer
          minimum: 1
          maximum: 5
        reviewText:
          type: string
        createdAt:
          type: string
          format: date-time

    ReviewInput:
      type: object
      required: [mediaType, externalId, title, rating, reviewText]
      properties:
        mediaType:
          type: string
          enum: [movie, book]
        externalId:
          type: string
        title:
          type: string
        coverImageUrl:
          type: ["string", "null"]
        year:
          type: ["string", "null"]
        rating:
          type: integer
          minimum: 1
          maximum: 5
        reviewText:
          type: string

    ReviewsPage:
      type: object
      required: [reviews, total, page, limit]
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    ReviewStats:
      type: object
      required: [totalReviews, totalMovieReviews, totalBookReviews, averageRating, topRated]
      properties:
        totalReviews:
          type: integer
        totalMovieReviews:
          type: integer
        totalBookReviews:
          type: integer
        averageRating:
          type: number
        topRated:
          type: array
          items:
            $ref: "#/components/schemas/Review"

    WatchlistItem:
      type: object
      required: [id, userId, mediaType, externalId, title, addedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        mediaType:
          type: string
          enum: [movie, book]
        externalId:
          type: string
        title:
          type: string
        coverImageUrl:
          type: ["string", "null"]
        addedAt:
          type: string
          format: date-time

    WatchlistInput:
      type: object
      required: [mediaType, externalId, title]
      properties:
        mediaType:
          type: string
          enum: [movie, book]
        externalId:
          type: string
        title:
          type: string
        coverImageUrl:
          type: ["string", "null"]

    SearchResult:
      type: object
      required: [externalId, title, mediaType]
      properties:
        externalId:
          type: string
        title:
          type: string
        mediaType:
          type: string
          enum: [movie, book]
        coverImageUrl:
          type: ["string", "null"]
        year:
          type: ["string", "null"]
        author:
          type: ["string", "null"]