# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Movie & Book Review API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  /** @minLength 2 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}