# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Movie & Book Review API
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewMediaType } from "./reviewMediaType";

export interface Review {
  id: number;
  userId: number;
  /** @nullable */
  username?: string | null;
  mediaType: ReviewMediaType;
  externalId: string;
  title: string;
  /** @nullable */
  coverImageUrl?: string | null;
  /** @nullable */
  year?: string | null;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  reviewText: string;
  createdAt: Date;
}