# lib/api-zod/src/generated/types/reviewInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Movie & Book Review API
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewInputMediaType } from "./reviewInputMediaType";

export interface ReviewInput {
  mediaType: ReviewInputMediaType;
  externalId: string;
  title: string;
  /** @nullable */
  coverImageUrl?: string | null;
  /** @nullable */
  year?: string | null;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  reviewText: string;
}