# lib/api-zod/src/generated/types/searchResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Movie & Book Review API
 * OpenAPI spec version: 0.1.0
 */
import type { SearchResultMediaType } from "./searchResultMediaType";

export interface SearchResult {
  externalId: string;
  title: string;
  mediaType: SearchResultMediaType;
  /** @nullable */
  coverImageUrl?: string | null;
  /** @nullable */
  year?: string | null;
  /** @nullable */
  author?: string | null;
}