# lib/api-zod/src/generated/types/watchlistInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Movie & Book Review API
 * OpenAPI spec version: 0.1.0
 */
import type { WatchlistInputMediaType } from "./watchlistInputMediaType";

export interface WatchlistInput {
  mediaType: WatchlistInputMediaType;
  externalId: string;
  title: string;
  /** @nullable */
  coverImageUrl?: string | null;
}