# lib/api-zod/src/generated/types/watchlistItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Movie & Book Review API
 * OpenAPI spec version: 0.1.0
 */
import type { WatchlistItemMediaType } from "./watchlistItemMediaType";

export interface WatchlistItem {
  id: number;
  userId: number;
  mediaType: WatchlistItemMediaType;
  externalId: string;
  title: string;
  /** @nullable */
  coverImageUrl?: string | null;
  addedAt: Date;
}